# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Climate Invest Optimizer API
  description: |
    API REST para la optimización de presupuestos de inmuebles orientada a la prevención 
    y mitigación de riesgos climáticos.
    
    ## Características principales
    
    - **Gestión de inmuebles (Shops)**: CRUD completo de tiendas con información de localización y riesgos
    - **Evaluación de riesgos climáticos**: Análisis de exposición, sensibilidad y probabilidad por cluster
    - **Medidas preventivas**: Catálogo de medidas con costos estimados y tipos
    - **Optimización de presupuesto**: Algoritmos (greedy, knapsack, weighted) para maximizar reducción de riesgo
    - **Dashboard**: Estadísticas y métricas generales
    
    ## Autenticación
    
    La API utiliza autenticación JWT. Incluye el token en el header `Authorization`:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Códigos de error
    
    | Código | Descripción |
    |--------|-------------|
    | 400 | Bad Request - Error de validación |
    | 401 | Unauthorized - Token inválido o expirado |
    | 403 | Forbidden - Sin permisos suficientes |
    | 404 | Not Found - Recurso no encontrado |
    | 409 | Conflict - Recurso duplicado |
    | 422 | Unprocessable Entity - Error de negocio |
    | 500 | Internal Server Error - Error del servidor |
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
paths:
...
  /clusters:
    get:
      tags:
        - clusters
      summary: Lista todos los clusters
      description: Obtiene la lista de todos los clusters geográficos
      operationId: listClusters
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de clusters
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Cluster'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /clusters/{id}:
    get:
      tags:
        - clusters
      summary: Obtener cluster por ID
      description: Retorna los detalles de un cluster incluyendo sus riesgos
      operationId: getCluster
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles del cluster
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ClusterWithRisks'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /measures:
    get:
      tags:
        - measures
      summary: Lista todas las medidas
      description: Obtiene el catálogo de medidas preventivas disponibles
      operationId: listMeasures
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          description: Filtrar por tipo de medida
          schema:
            type: string
            enum: [natural, material, immaterial]
      responses:
        '200':
          description: Lista de medidas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Measure'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /risks:
    get:
      tags:
        - risks
      summary: Lista todos los riesgos
      description: Obtiene la lista de todos los riesgos climáticos
      operationId: listRisks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de riesgos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Risk'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /optimization/budget:
    post:
      tags:
        - optimization
      summary: Optimizar presupuesto
      description: |
        Calcula la distribución óptima de medidas preventivas para un presupuesto dado.
        
        ## Estrategias disponibles
        
        - **greedy**: Selecciona medidas por mejor ratio costo-beneficio. Rápido y efectivo.
        - **knapsack**: Programación dinámica para solución óptima. Más preciso pero costoso.
        - **weighted**: Considera prioridades de riesgos definidas por el usuario.
        
      operationId: optimizeBudget
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizeBudgetRequest'
            examples:
              basic:
                summary: Optimización básica
                value:
                  shop_ids: [1, 2, 3]
                  max_budget: 50000
                  strategy: greedy
              withPriorities:
                summary: Con prioridades de riesgo
                value:
                  shop_ids: [1, 2, 3]
                  max_budget: 100000
                  strategy: weighted
                  risk_priorities: [1, 5, 8]
      responses:
        '200':
          description: Resultado de la optimización
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OptimizationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /dashboard/stats:
    get:
      tags:
        - dashboard
      summary: Estadísticas del dashboard
      description: Retorna métricas generales para el dashboard principal
      operationId: getDashboardStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Estadísticas del dashboard
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DashboardStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /auth/login
...
  responses:
    BadRequest:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: "El campo 'location' es requerido"
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Token de autenticación requerido
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Tienda no encontrada
...
    UnprocessableEntity:
      description: Error de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INSUFFICIENT_BUDGET
              message: El presupuesto es insuficiente para cualquier medida
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operación exitosa
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: ERROR_CODE
            message:
              type: string
              example: Descripción del error
...
    Cluster:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Mediterranean Coast
        utm_north:
          type: number
          format: float
          example: 41.0
        utm_east:
          type: number
          format: float
          example: 2.0
    ClusterWithRisks:
      allOf:
        - $ref: '#/components/schemas/Cluster'
        - type: object
          properties:
            risks:
              type: array
              items:
                $ref: '#/components/schemas/RiskDetail'
    Risk:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Inundación costera/fluvial/pluvial
        svg:
          type: string
          nullable: true
    RiskDetail:
      allOf:
        - $ref: '#/components/schemas/Risk'
        - type: object
          properties:
            exposure:
              type: string
              enum: [very_low, low, medium, high, very_high]
              example: medium
            sensitivity:
              type: string
              enum: [very_low, low, medium, high, very_high]
              example: high
            consequence:
              type: string
              enum: [very_low, low, medium, high, very_high]
              example: high
            probability:
              type: string
              enum: [very_low, low, medium, high, very_high]
              example: medium
            risk_score:
              type: number
              format: float
              example: 0.42
    Measure:
      type: object
      properties:
        name:
          type: string
          example: Sistema de drenaje perimetral
        estimated_cost:
          type: number
          format: float
          example: 4500
        type:
          type: string
          enum: [natural, material, immaterial]
          example: material
...
    OptimizeBudgetRequest:
      type: object
      required:
        - shop_ids
        - max_budget
      properties:
        shop_ids:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
          example: [1, 2, 3]
        max_budget:
          type: number
          format: float
          minimum: 0
          example: 50000
        strategy:
          type: string
          enum: [greedy, knapsack, weighted]
          default: greedy
          description: |
            - **greedy**: Rápido, buen balance costo-beneficio
            - **knapsack**: Óptimo pero más lento
            - **weighted**: Considera prioridades de riesgo
        risk_priorities:
          type: array
          items:
            type: integer
            format: int64
          description: IDs de riesgos prioritarios (solo para strategy=weighted)
          example: [1, 5]
    OptimizationResult:
      type: object
      properties:
        total_cost:
          type: number
          format: float
          example: 45500
        remaining_budget:
          type: number
          format: float
          example: 4500
        total_risk_reduction:
          type: number
          format: float
          description: Porcentaje de reducción de riesgo
          example: 35.5
        recommended_measures:
          type: array
          items:
            $ref: '#/components/schemas/RecommendedMeasure'
        shop_recommendations:
          type: array
          items:
            $ref: '#/components/schemas/ShopRecommendation'
        strategy_used:
          type: string
          example: greedy
        metrics:
          $ref: '#/components/schemas/OptimizationMetrics'
    RecommendedMeasure:
      type: object
      properties:
        measure:
          $ref: '#/components/schemas/Measure'
        priority:
          type: integer
          example: 1
        risk_reduction_percentage:
          type: number
          format: float
          example: 12.5
        cost_efficiency_score:
          type: number
          format: float
          example: 0.0028
        affected_risks:
          type: array
          items:
            type: string
          example: ["Inundación costera/fluvial/pluvial", "Precipitaciones intensas"]
        justification:
          type: string
          example: Medida con alta eficiencia que reduce riesgos relacionados con eventos climáticos
    ShopRecommendation:
      type: object
      properties:
        shop_id:
          type: integer
          format: int64
        shop_location:
          type: string
        current_risk:
          type: number
          format: float
        projected_risk:
          type: number
          format: float
        measures:
          type: array
          items:
            $ref: '#/components/schemas/RecommendedMeasure'
        estimated_investment:
          type: number
          format: float
    OptimizationMetrics:
      type: object
      properties:
        budget_utilization_percentage:
          type: number
          format: float
          example: 91.0
        average_risk_reduction:
          type: number
          format: float
          example: 15.5
        estimated_roi:
          type: number
          format: float
          example: 2.5
        processing_time_ms:
          type: integer
          format: int64
          example: 45
    DashboardStats:
      type: object
      properties:
        total_shops:
          type: integer
          format: int64
          example: 150
        total_clusters:
          type: integer
          format: int64
          example: 12
        average_risk:
          type: number
          format: float
          example: 0.45
        high_risk_shops:
          type: integer
          format: int64
          example: 23
        total_measures:
          type: integer
          format: int64
          example: 40
        applied_measures:
          type: integer
          format: int64
          example: 185
        total_investment:
          type: number
          format: float
          example: 450000
        coverage_percentage:
          type: number
          format: float
          example: 65.5